requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
groq
langchain-groq
sentence_transformers
langchain-huggingface